requirements.txt
azure-functions
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
pytest==8.4.0
requests==2.31.0 